Login changes made on top of the refactored code
