+++ Frontend/channelx/package.json
@@ -6,7 +6,7 @@
     "bootstrap": "^4.3.1",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.14.0",
-    "firebase": "^7.0.0",
+    "firebase": "^7.1.0",
     "react": "^16.10.1",
     "react-dom": "^16.10.1",
     "react-router": "^5.0.1",
