+++ Frontend/channelx/src/App.js
@@ -8,6 +8,7 @@ import React,{Component} from 'react';
 import './App.css';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import Signup from './components/Signup/Signup';
+import Login from './components/Login/Login';
 import Landing from './components/Landing/Landing';
 import EmailSent from './components/VerificationEmailPage/EmailSent';
 import Fire from "./config/Fire";
@@ -45,6 +46,7 @@ class App extends Component {
                     <div className="App">
                         <Route path={ROUTES.LANDING} exact component={Landing}/>
                         <Route path={ROUTES.SIGN_UP}  exact component={Signup}/>
+                        <Route path={ROUTES.SIGN_IN} exact component={Login}/>
                         <Route path={ROUTES.HOME}  exact component={Home}/>
                         <Route path={ROUTES.EMAIL_SENT}  exact component={EmailSent}/>
                     </div>
