--- Frontend/channelx/src/Landing/Landing.css
@@ -1,54 +1,0 @@
-/*
-Description: styling code for landing page
-Authors: Manisha Miriyala
-Date: 9/20/2019
-*/
-
-+/* Colors */
-    /*
-    dark blue: #476A6F
-    light blue: #258EA6
-    dark green: #519E8A
-    medium green: #7EB09B
-    light green: #A3C4BC
-    */
-
-.wrapper {
-    height: 100vh;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    background-color: #7678ED;
-}
-
-.Register {
-    width: 100%;
-    flex-direction: column;
-    align-items: center;
-}
-
-.Register button {
-    color: #000000;
-    border: 2px solid #fff;
-    width: 30%;
-    margin-top: 1em;
-    padding: 8px 0px;
-    font-size: 1em;
-    font-weight: lighter;
-    letter-spacing: 1px;
-    margin-bottom: 0.25em;
-    border-radius: 15px;
-}
-
-.Register button:hover {
-    color: #519e8a;
-    background-color: #fff;
-    border: 2px solid #519e8a;
-}
-
-.Register small {
-    color: #999;
-    font-weight: lighter;
-}
