+++ Frontend/channelx/src/constants/keys/keys.js
@@ -1,0 +1,14 @@
+/*
+Description: Saving Firebase Keys so that no one commits them
+Author: Sami
+Date: 10/02/2019
+*/
+
+export const API_KEY = 'AIzaSyBTi6-MYDO-skqiSeuEJrlpaqe6URhUBtc';
+export const AUTH_DOMAIN = 'channelx-714c6.firebaseapp.com';
+export const DATABASE_URL = 'https://channelx-714c6.firebaseio.com';
+export const PROJECT_ID = 'channelx-714c6';
+export const STORAGE_BUCKET = '';
+export const MESSAGING_SENDER_ID = '697701922217';
+export const APP_ID = '1:697701922217:web:03e8c2f4961ea7228f00b8';
+export const MEASUREMENTS_ID = 'G-7SJGHG1M48';
