+++ Frontend/channelx/src/constants/routes.js
@@ -1,6 +1,6 @@
 export const LANDING = '/';
 export const SIGN_UP = '/signup';
-export const SIGN_IN = '/signin';
+export const SIGN_IN = '/login';
 export const HOME = '/home';
 export const EMAIL_SENT = '/emailsent'
 export const ACCOUNT = '/account';
